requirements.txt

# ========================================
# Pilot Atendimento MVE — Dependências
# ========================================
# Versão: v1.1
# Python: 3.11+

# ========================================
# FRAMEWORK WEB
# ========================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ========================================
# DATABASE — PostgreSQL (async)
# ========================================
asyncpg>=0.29.0

# ========================================
# LLM — Google Gemini
# ========================================
google-genai>=0.3.0

# ========================================
# VECTOR STORE — ChromaDB
# ========================================
chromadb>=0.4.22

# ========================================
# UTILIDADES
# ========================================
httpx>=0.26.0
python-dotenv>=1.0.0

# ========================================
# OBSERVABILITY
# ========================================
prometheus-fastapi-instrumentator>=7.0.0

# ========================================
# SEGURANÇA
# ========================================
slowapi>=0.1.9  # Rate limiting
tenacity>=8.2.3  # Retry com backoff exponencial
boto3>=1.28.0  # AWS Secrets Manager (opcional)
bcrypt>=4.1.0  # Hash de senhas
pyjwt>=2.8.0  # JWT authentication

# ========================================
# SCRAPING
# ========================================
playwright>=1.40.0
